backend/requirements.txt
fastapi
slowapi
uvicorn[standard]
openai
weaviate-client
# python-multipart
python-dotenv
redis
starlette
json_log_formatter